Split voice types
