+++ packages/types/src/index.ts
@@ -20,6 +20,7 @@ export * from './discord/stageInstances.js'
 export * from './discord/stickers.js'
 export * from './discord/subscriptions.js'
 export * from './discord/users.js'
+export * from './discord/voice.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
 export * from './shared.js'
